# accounting-api/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting Service API
  description: A comprehensive accounting API similar to QuickBooks
  version: 1.0.0
  contact:
    email: api@accounting-service.com
servers:
  - url: https://api.accounting-service.com/v1
security:
  - bearerAuth: []
  - oauth2: [read, write]
...
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List all accounts (Chart of Accounts)
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AccountType'
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
    post:
      tags: [Accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /accounts/{accountId}:
    parameters:
      - $ref: '#/components/parameters/accountId'
    get:
      tags: [Accounts]
      summary: Get account by ID
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    patch:
      tags: [Accounts]
      summary: Update an account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: Account updated
    delete:
      tags: [Accounts]
      summary: Delete an account
      responses:
        '204':
          description: Account deleted
  /accounts/{accountId}/transactions:
    get:
      tags: [Accounts]
      summary: Get transactions for an account
      parameters:
        - $ref: '#/components/parameters/accountId'
        - $ref: '#/components/parameters/startDate'
        - $ref: '#/components/parameters/endDate'
      responses:
        '200':
          description: Account transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'

  # ============ CUSTOMERS ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Read access
            write: Write access
            admin: Admin access
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
        maximum: 100
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
    startDate:
      name: startDate
      in: query
      schema:
        type: string
        format: date
    endDate:
      name: endDate
      in: query
      schema:
        type: string
        format: date
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    AccountType:
      type: string
      enum:
        - asset
        - liability
        - equity
        - income
        - expense
...
    Money:
      type: object
      properties:
        amount:
          type: number
          format: decimal
        currency:
          type: string
          default: USD

    # === Address ===
...
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        accountNumber:
          type: string
        description:
          type: string
        balance:
          $ref: '#/components/schemas/Money'
        parentAccountId:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AccountCreate:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        accountNumber:
          type: string
        description:
          type: string
        parentAccountId:
          type: string
        openingBalance:
          $ref: '#/components/schemas/Money'
    AccountUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    AccountList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # === Customer ===
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean

    # === Placeholder schemas (incomplete in spec) ===
    TransactionList:
      type: object
...